requirements.txt
docopt==0.6.2
pyredmine==0.2.4
python-dateutil==2.4.2
python-harvest==1.0.4
requests==2.9.1
six==1.10.0
texttable==0.8.2
tqdm==3.8.0
wheel==0.24.0